Update helpers tests
